rename: 학생 인증 페이지 경로 수정 (#9)
