+++ src/pages/signup/verify.tsx
@@ -1,6 +1,6 @@
 import React, { FormEvent, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { ROUTES } from '../constants';
+import { ROUTES } from '../../constants';
 import axios from 'axios';
 
 export default function SignupVerify() {
